moved mapping of stylee
