+++ amundsen_application/static/js/config/config-utils.ts
@@ -90,16 +90,6 @@ export function getBadgeConfig(badgeName: string): BadgeStyleConfig {
   };
 }
 
-/*
- * maps badge type to a badge style
- */
-export function mapBadgeStyle(badgeType: string): BadgeStyle {
-  if (badgeType === 'negative') return BadgeStyle.DANGER;
-  if (badgeType === 'positive') return BadgeStyle.SUCCESS;
-  if (badgeType === 'warning') return BadgeStyle.WARNING;
-  return BadgeStyle.DEFAULT;
-}
-
 /**
  * Returns whether or not feedback features should be enabled
  */
